+++ flask_rest_api/error_handler.py
@@ -1,7 +1,7 @@
 """Exception handler"""
 
-from werkzeug.exceptions import HTTPException, InternalServerError
-from flask import jsonify, current_app
+from werkzeug.exceptions import HTTPException
+from flask import jsonify
 
 
 class ErrorHandlerMixin:
@@ -10,7 +10,7 @@ class ErrorHandlerMixin:
     def _register_error_handlers(self):
         """Register error handlers in Flask app
 
-        This method registers a default error handler for HTTPException.
+        This method registers a default error handler for ``HTTPException``.
         """
         self._app.register_error_handler(
             HTTPException, self.handle_http_exception)
@@ -18,13 +18,13 @@ class ErrorHandlerMixin:
     def handle_http_exception(self, error):
         """Return a JSON response containing a description of the error
 
-        This method is registered at app init to handle `HTTPException`.
+        This method is registered at app init to handle ``HTTPException``.
 
-        - When `abort` is called in the code, this triggers a `HTTPException`,
-          and Flask calls this handler to generate a better response.
+        - When ``abort`` is called in the code, an ``HTTPException`` is
+          triggered and Flask calls this handler.
 
-        - When an exception is not caught in a view, Flask automatically calls
-          the handler registered for error 500.
+        - When an exception is not caught in a view, Flask makes it an
+          ``InternalServerError`` and calls this handler.
 
         flask_rest_api republishes webargs's
         :func:`abort <webargs.flaskparser.abort>`. This ``abort`` allows the
@@ -34,33 +34,12 @@ class ErrorHandlerMixin:
         Extra information expected by this handler:
 
         - `message` (``str``): a comment
-        - `errors` (``dict``): errors, typically validation issues on a form
+        - `errors` (``dict``): errors, typically validation errors in
+            parameters and request body
         - `headers` (``dict``): additional headers
-
-        If the error is an ``HTTPException`` (typically if it was triggered by
-        ``abort``), this handler logs it with `INF0` level. Otherwise, it is an
-        unhandled exception and it is already logged as `ERROR` by Flask.
         """
-        # TODO: use an error Schema
-        # TODO: add a parameter to enable/disable logging?
-
-        # If error is not a HTTPException, then it is an unhandled exception.
-        # Make it a 500 (InternalServerError) and don't log.
-        do_log = True
-        if not isinstance(error, HTTPException):
-            error = InternalServerError()
-            do_log = False
-
-        payload, headers = self._prepare_error_response_content(error)
-        if do_log:
-            self._log_error(error, payload)
-        return jsonify(payload), error.code, headers
-
-    @staticmethod
-    def _prepare_error_response_content(error):
-        """Build payload and headers from error"""
         headers = {}
-        payload = {'status': str(error), }
+        payload = {'code': error.code, 'status': error.name}
 
         # Get additional info passed as kwargs when calling abort
         # data may not exist if
@@ -83,12 +62,4 @@ class ErrorHandlerMixin:
             if 'headers' in data:
                 headers = data['headers']
 
-        return payload, headers
-
-    @staticmethod
-    def _log_error(error, payload):
-        """Log error as INFO, including payload content"""
-        log_string_content = [str(error.code), ]
-        log_string_content.extend([
-            str(payload[k]) for k in ('message', 'errors') if k in payload])
-        current_app.logger.info(' '.join(log_string_content))
+        return jsonify(payload), error.code, headers